small refactor & memoization
