+++ project/src/components/index.ts
@@ -18,13 +18,14 @@ export { default as Main } from './main/main';
 export { default as MainEmpty } from './main-empty/main-empty';
 export { default as Map } from './map/map';
 export { default as NotFoundScreen } from './not-found-screen/not-found-screen';
-export { default as Offer } from './offer/offer';
 export { default as OfferCard } from './offer-card/offer-card';
 export { default as OfferGood } from './offer-good/offer-good';
 export { default as OfferList } from './offer-list/offer-list';
 export { default as OfferLoadingScreen } from './offer-loading-screen/offer-loading-screen';
 export { default as OfferNotLogged } from './offer-not-logged/offer-not-logged';
+export { default as OfferPage } from './offer-page/offer-page';
 export { default as OfferPromoImg } from './offer-promo-img/offer-promo-img';
+export { default as Offers } from './offers/offers';
 export { default as PrivateRoute } from './private-route/private-route';
 export { default as SortingItem } from './sorting-item/sorting-item';
 export { default as SortingOptions } from './sorting-options/sorting-options';
