+++ project/src/components/offer-page/const.ts
