+++ project/src/components/offer-page/offer-page.tsx
