+++ project/src/components/offer-page/types.ts
