+++ project/src/consts/offer.ts
@@ -1,4 +1,4 @@
-import type { ActiveOffer } from '../../types';
+import type { ActiveOffer } from '../types';
 
 export const initalActiveOffer: ActiveOffer = {
   id: null,
